Pal Games don`t work
Greetings. Installed xstation in console 1001 ntsc. I use it to connect RGB Scart to a CRT TV. NTSC images work fine. But PAL images don't work. In most cases, this is a dark screen, or just sound without a picture. I think the xstation is installed correctly, because that's the only problem. If you switch the picture mode to PAL in the menu in the menu, then there will be a dark screen and the firmware has to be updated again. I know that xstation should remove regional restrictions, so this worries me. Why is that? If I buy a PAL set-top box, do I understand correctly that the problem will be only in black and white, which will be solved by RGB scart? Why doesn't the reverse order work for me? Thank you.